fix: 멀티 모듈 초기 구성
