--- project/cheolhyun-onboarding/Survey/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "Survey"
