--- project/cheolhyun-onboarding/Survey/src/main/java/com/onboarding/survey/SurveyApplication.java
@@ -1,13 +1,0 @@
-package com.onboarding.survey;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SurveyApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SurveyApplication.class, args);
-    }
-
-}
