--- project/cheolhyun-onboarding/Survey/src/main/java/com/onboarding/survey/domain/model/QuestionType.java
@@ -1,8 +1,0 @@
-package com.onboarding.survey.domain.model;
-
-public enum QuestionType {
-    SHORT_TEXT,
-    LONG_TEXT,
-    SINGLE_CHOICE,
-    MULTI_CHOICE
-}
