--- project/cheolhyun-onboarding/Survey/src/test/java/com/onboarding/survey/SurveyApplicationTests.java
@@ -1,12 +1,0 @@
-package com.onboarding.survey;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SurveyApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-}
