--- project/cheolhyun-onboarding/Survey/src/test/java/com/onboarding/survey/domain/model/OptionsTest.java
@@ -1,20 +1,0 @@
-package com.onboarding.survey.domain.model;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.*;
-
-class OptionsTest {
-
-    @Test
-    @DisplayName("옵션 넣기")
-    void add() {
-        Options result1 = new Options("Option1");
-        Options result2 = new Options();
-
-        result2.add("Option1");
-
-        assertThat(result1).isEqualTo(result2);
-    }
-}
