+++ project/cheolhyun-onboarding/build.gradle.kts
@@ -1,0 +1,24 @@
+plugins {
+    java
+}
+
+group = "com.onboarding"
+version = "0.0.1-SNAPSHOT"
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+}
+
+tasks.withType<Test> {
+    useJUnitPlatform()
+}
