+++ project/cheolhyun-onboarding/core/src/main/java/com/onboarding/model/QuestionType.java
@@ -1,0 +1,8 @@
+package com.onboarding.model;
+
+public enum QuestionType {
+    SHORT_TEXT,
+    LONG_TEXT,
+    SINGLE_CHOICE,
+    MULTI_CHOICE
+}
