+++ project/cheolhyun-onboarding/core/src/test/java/survey/domain/model/SurveyTest.java
@@ -1,0 +1,19 @@
+package survey.domain.model;
+
+import com.onboarding.model.Survey;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.*;
+
+class SurveyTest {
+    @Test
+    @DisplayName("null과 빈값 유효성 검증")
+    void validEmpty() {
+        assertAll(
+                () -> assertThatThrownBy(() -> new Survey("")).isInstanceOf(IllegalArgumentException.class),
+                () -> assertThatThrownBy(() -> new Survey(null)).isInstanceOf(IllegalArgumentException.class)
+        );
+    }
+}
