+++ project/cheolhyun-onboarding/settings.gradle.kts
@@ -1,0 +1,6 @@
+rootProject.name = "cheolhyun-onboarding"
+
+include("core")
+include("app")
+include("core")
+include("infrastructure")
